Add error handling callback to receiver type
Users can now register an error handling callback. This procedure is
called if an error occurs when receiving or processing data. Setting the
callback's stop flag to True tells the receiver to stop listening for
data and terminate (e.g. when the error is fatal).

The error handling callback must be registered before instructing the
receiver to listen for data.