Suppress SIGPIPE signal for send operations
Suppress SIGPIPE signal upon writing attempts on disconnected sockets.
The POSIX standard requires that a SIGPIPE signal is generated when
attempting to send on a stream-oriented socket that is no longer
connected.

However, directly handling signals in an Ada program should be avoided
where possible.

Setting the MSG_NOSIGNAL flag of the send and sendto functions will
suppress this signal. The EPIPE error is still returned by those
functions in that case.