Implement socket shutdown function with associated test
The POSIX shutdown function provides finer control on how to disable
open socket communications.  Whereas the close function will disable
both sending and receiving operations on the socket, the shutdown
function can shutdown either side of the full-duplex connection as well
as both sides (which is equivalent to the close function).